fixing import issues, created a generalized utils module
