+++ app/facial_recognition/utils.py
@@ -1,4 +1,4 @@
-# Utilities module to hold helper functions for recognition.py
+# General utilities module to hold helper functions
 
 import cv2
 import face_recognition as fr
@@ -7,7 +7,9 @@ import json
 from datetime import datetime
 import csv
 import numpy as np
-from event_controller import end_event
+import serial
+import serial.tools.list_ports
+from threading import Event
 
 facial_path = os.path.dirname(os.path.abspath(__file__))                            # Path to facial_recognition/
 app_path = os.path.dirname(facial_path)                                             # Path to app/ 
@@ -15,8 +17,17 @@ attendance_path = os.path.join(app_path, 'attendance')                          
 
 supported_extensions = ('.jpg', '.jpeg', '.png')
 
-def encode_image(file):
 
+# Event to control RFID polling and facial recognition
+rfid_event = Event()
+rfid_event.set()  # Start with RFID scanning enabled
+
+# Event to control when to end attendance
+end_event = Event()
+end_event.clear()   # start with end_event cleared
+
+
+def encode_image(file):
     # if invalid extension type
     if not file.filename.endswith(supported_extensions):
         raise ValueError(f'Invalid file type! Supported types are {supported_extensions}')
@@ -63,6 +74,57 @@ def display_camera(cap):
     cap.release()
     cv2.destroyAllWindows()
     
+
+
+def poll_rfid(cap, users):
+    try:
+        ser = serial.Serial('/dev/tty.usbserial-1410', 9600, timeout=1) # make it dynamically choose the serial port depending on the OS
+        time.sleep(2)  # Allow time for the connection to establish
+        print(f"Connected to /dev/tty.usbserial-1410")
+
+        while not end_event.is_set():
+            rfid_event.wait()
+            # rfid_event.wait()  # Only proceed if rfid_event is set
+
+            if ser.in_waiting > 0:
+                rfid_data = ser.readline().decode('utf-8').strip()
+                print(f"RFID Tag: {rfid_data}")
+                rfid_event.clear()
+
+                # check if this RFID is valid for the event
+                if is_valid_for_event(rfid_data, users):
+                    print('Valid RFID for event')
+                    verified = start_facial_recognition(cap, rfid_data, users)
+
+                    if verified:
+                        print('Verified user! Marking attendance...')
+                        mark_attendance(verified)
+                    else:
+                        print('Face not recognized or timed out. Please rescan RFID and try again.')
+
+                else:
+                    print('RFID is not associated with this event.')
+                
+                rfid_event.set()
+            
+            time.sleep(0.1)
+
+    except serial.SerialException as e:
+        print(f"Serial Error: {e}")
+    except PermissionError as e:
+        print(f"Permission Error: {e}")
+    except Exception as e:
+        print(f"Unexpected Error: {e}")
+    finally:
+        ser.close()
+        end_event.set()
+        print("Serial port closed.")
+
+
+def is_valid_for_event(rfid_data, users):
+    rfid_list = [user.rfid for user in users]
+    return rfid_data in rfid_list
+
 
 def mark_attendance(user):
     current_datetime = datetime.now().strftime("%B %d, %Y %I:%M:%S %p")
